Getdetails command with public key, address, native token ref
